main.dart に定義されていたWidgetを別ファイルに分割
