+++ project/bodquest_2023/lib/presentation/page/login_page.dart
@@ -1,11 +1,12 @@
 // ignore_for_file: unused_field
 
 import 'package:bodquest_2023/core/exception/firebasse_auth_exception.dart';
-import 'package:bodquest_2023/main.dart';
 import 'package:bodquest_2023/presentation/page/registration_page.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+
+import 'myhome_page.dart';
 
 class LogInPage extends StatefulWidget {
   @override
