+++ project/bodquest_2023/lib/presentation/page/myapp_page.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+import 'login_page.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Bodquest 2023',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: LogInPage(), //const MyHomePage(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
