+++ project/bodquest_2023/lib/presentation/page/registration_page.dart
@@ -1,8 +1,9 @@
 import 'package:bodquest_2023/core/exception/firebasse_auth_exception.dart';
-import 'package:bodquest_2023/main.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+
+import 'myhome_page.dart';
 
 // アカウント登録ページ
 class RegistrationPage extends StatefulWidget {
@@ -51,7 +52,8 @@ class _RegistrationPageState extends State<RegistrationPage> {
                 decoration: InputDecoration(labelText: "パスワード（8～20文字）"),
                 obscureText: true, // パスワードが見えないようにする
                 maxLength: 20, // 入力可能な文字数
-                maxLengthEnforcement: MaxLengthEnforcement.enforced, // 入力可能な文字数の制限を超える場合の挙動の制御
+                maxLengthEnforcement:
+                    MaxLengthEnforcement.enforced, // 入力可能な文字数の制限を超える場合の挙動の制御
                 onChanged: (String value) {
                   if (value.length >= 8) {
                     _password = value;
@@ -66,14 +68,15 @@ class _RegistrationPageState extends State<RegistrationPage> {
             // 登録失敗時のエラーメッセージ
             Padding(
               padding: EdgeInsets.fromLTRB(20.0, 0, 20.0, 5.0),
-              child: Text(_infoText,style: TextStyle(color: Colors.red)),
+              child: Text(_infoText, style: TextStyle(color: Colors.red)),
             ),
 
             ButtonTheme(
               minWidth: 350.0,
               // height: 100.0,
               child: ElevatedButton(
-                child: Text('登録', style: TextStyle(fontWeight: FontWeight.bold)),
+                child:
+                    Text('登録', style: TextStyle(fontWeight: FontWeight.bold)),
                 onPressed: () async {
                   if (_validPassword) {
                     try {
